# api-specification-singulary.yaml
openapi: 3.0.3
info:
  title: Singulary API
  description: |
    API completa para o aplicativo Singulary - Ressonador Quântico Pessoal (RQP).
    
    Esta API fornece endpoints para gerenciamento de usuários, presets de frequências,
    sessões de meditação/terapia sonora e integração futura com hardware.
    
    ## Funcionalidades Principais
    - Autenticação JWT
    - Gerenciamento de presets (padrão e personalizados)
    - Registro e análise de sessões
    - Exportação de dados em CSV
    - Estatísticas e relatórios
    - Preparação para integração com hardware
    
    ## Segurança
    - Autenticação via JWT tokens
    - Validação de dados de entrada
    - Rate limiting (implementação futura)
    - CORS configurado para frontend
    
    ## Versioning
    Esta é a versão 1.0 da API, compatível com o MVP do Singulary.
    
  version: 1.0.0
  contact:
    name: Equipe Singulary
    email: api@singulary.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desenvolvimento local
  - url: https://api.singulary.app/v1
    description: Servidor de produção (futuro)
...
paths:
...
  /sessions:
    get:
      tags:
        - sessions
      summary: Listar sessões do usuário
      description: |
        Obter lista de sessões do usuário com filtros opcionais.
        Inclui informações detalhadas de cada sessão e preset utilizado.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: true
          description: ID do usuário
          schema:
            type: string
            example: user-abc123
        - name: dimension
          in: query
          description: Filtrar por dimensão do preset
          schema:
            type: string
            enum: [3D, 4D, 5D, Personalizado]
            example: 3D
        - name: start_date
          in: query
          description: Data de início do filtro (ISO 8601)
          schema:
            type: string
            format: date-time
            example: 2025-08-01T00:00:00Z
        - name: end_date
          in: query
          description: Data de fim do filtro (ISO 8601)
          schema:
            type: string
            format: date-time
            example: 2025-08-31T23:59:59Z
        - name: limit
          in: query
          description: Número máximo de sessões a retornar
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
            example: 20
      responses:
        '200':
          description: Lista de sessões
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - sessions
      summary: Criar nova sessão
      description: Registrar o início de uma nova sessão de meditação/terapia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - presetId
                - startTime
              properties:
                userId:
                  type: string
                  example: user-abc123
                  description: ID do usuário
                presetId:
                  type: string
                  example: 3d-focus
                  description: ID do preset utilizado
                startTime:
                  type: string
                  format: date-time
                  example: 2025-08-10T14:30:00Z
                  description: Horário de início da sessão
                endTime:
                  type: string
                  format: date-time
                  example: 2025-08-10T14:50:00Z
                  description: Horário de fim (opcional, para sessões já finalizadas)
                duration:
                  type: integer
                  example: 1200
                  description: Duração em segundos (opcional)
                preMoodScore:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 6
                  description: Avaliação de humor antes da sessão (1-10)
                postMoodScore:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 8
                  description: Avaliação de humor após a sessão (1-10)
                notes:
                  type: string
                  example: Me senti mais concentrado e relaxado
                  description: Notas pessoais sobre a sessão
                deviceModel:
                  type: string
                  example: iPhone 13
                  description: Modelo do dispositivo utilizado
                headphonesUsed:
                  type: boolean
                  example: true
                  description: Se foram utilizados fones de ouvido
                roomDark:
                  type: boolean
                  example: false
                  description: Se o ambiente estava escuro
      responses:
        '201':
          description: Sessão criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/{sessionId}:
    get:
      tags:
        - sessions
      summary: Obter sessão específica
      description: Obter detalhes de uma sessão específica
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID da sessão
          schema:
            type: string
            example: session-abc123def456
      responses:
        '200':
          description: Detalhes da sessão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - sessions
      summary: Atualizar sessão
      description: Atualizar uma sessão (principalmente para finalizar)
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID da sessão
          schema:
            type: string
            example: session-abc123def456
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                endTime:
                  type: string
                  format: date-time
                  example: 2025-08-10T14:50:00Z
                  description: Horário de fim da sessão
                duration:
                  type: integer
                  example: 1200
                  description: Duração total em segundos
                postMoodScore:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 8
                  description: Avaliação de humor após a sessão
                notes:
                  type: string
                  example: Experiência muito relaxante
                  description: Notas sobre a sessão
                headphonesUsed:
                  type: boolean
                  example: true
                  description: Se foram utilizados fones de ouvido
                roomDark:
                  type: boolean
                  example: false
                  description: Se o ambiente estava escuro
      responses:
        '200':
          description: Sessão atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/{sessionId}/export:
    get:
      tags:
        - sessions
      summary: Exportar sessão como CSV
      description: Exportar uma sessão específica em formato CSV
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID da sessão
          schema:
            type: string
            example: session-abc123def456
      responses:
        '200':
          description: Arquivo CSV da sessão
          content:
            text/csv:
              schema:
                type: string
                example: |
                  session_id,user_id,date_utc,timezone,preset_name,dimension_tag,base_frequency_hz,beat_frequency_hz,light_frequency_hz,duration_seconds,headphones_used,room_dark,pre_mood_score,post_mood_score,notes_text,device_model,firmware_version,recording_path,session_hash
                  session-abc123,user-abc123,2025-08-10T14:30:00Z,UTC,Foco 3D,3D,528,40,10,1200,yes,no,6,8,"Me senti concentrado",iPhone 13,1.0.0,,hash_session-abc123
          headers:
            Content-Disposition:
              description: Nome do arquivo para download
              schema:
                type: string
                example: attachment; filename=session_abc123.csv
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/export:
    get:
      tags:
        - sessions
      summary: Exportar múltiplas sessões como CSV
      description: Exportar sessões do usuário em formato CSV com filtros opcionais
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: true
          description: ID do usuário
          schema:
            type: string
            example: user-abc123
        - name: dimension
          in: query
          description: Filtrar por dimensão
          schema:
            type: string
            enum: [3D, 4D, 5D, Personalizado]
            example: 3D
        - name: start_date
          in: query
          description: Data de início do filtro
          schema:
            type: string
            format: date-time
            example: 2025-08-01T00:00:00Z
        - name: end_date
          in: query
          description: Data de fim do filtro
          schema:
            type: string
            format: date-time
            example: 2025-08-31T23:59:59Z
      responses:
        '200':
          description: Arquivo CSV com múltiplas sessões
          content:
            text/csv:
              schema:
                type: string
          headers:
            Content-Disposition:
              description: Nome do arquivo para download
              schema:
                type: string
                example: attachment; filename=singulary_sessions_user-abc123.csv
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/stats:
    get:
      tags:
        - sessions
      summary: Obter estatísticas das sessões
      description: |
        Obter estatísticas agregadas das sessões do usuário, incluindo:
        - Total de sessões e minutos
        - Melhoria média de humor
        - Distribuição por dimensão e categoria
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: true
          description: ID do usuário
          schema:
            type: string
            example: user-abc123
      responses:
        '200':
          description: Estatísticas das sessões
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSessions:
                    type: integer
                    example: 25
                    description: Total de sessões realizadas
                  totalMinutes:
                    type: integer
                    example: 520
                    description: Total de minutos de sessões
                  avgMoodImprovement:
                    type: number
                    example: 1.8
                    description: Melhoria média de humor (diferença pós - pré)
                  dimensionStats:
                    type: object
                    example:
                      "3D": 15
                      "4D": 8
                      "5D": 2
                    description: Número de sessões por dimensão
                  categoryStats:
                    type: object
                    example:
                      "Foco": 10
                      "Relaxamento": 8
                      "Meditação": 5
                      "Sono": 2
                    description: Número de sessões por categoria
                  moodSessionsCount:
                    type: integer
                    example: 20
                    description: Número de sessões com avaliação de humor
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # USUÁRIOS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido através do endpoint de login.
        Formato: `Bearer <token>`
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          example: session-abc123def456
          description: ID único da sessão
        userId:
          type: string
          example: user-abc123
          description: ID do usuário
        presetId:
          type: string
          example: 3d-focus
          description: ID do preset utilizado
        startTime:
          type: string
          format: date-time
          example: 2025-08-10T14:30:00Z
          description: Horário de início da sessão
        endTime:
          type: string
          format: date-time
          nullable: true
          example: 2025-08-10T14:50:00Z
          description: Horário de fim da sessão
        duration:
          type: integer
          nullable: true
          example: 1200
          description: Duração em segundos
        preMoodScore:
          type: integer
          nullable: true
          minimum: 1
          maximum: 10
          example: 6
          description: Avaliação de humor antes da sessão (1-10)
        postMoodScore:
          type: integer
          nullable: true
          minimum: 1
          maximum: 10
          example: 8
          description: Avaliação de humor após a sessão (1-10)
        notes:
          type: string
          nullable: true
          example: Me senti mais concentrado e relaxado
          description: Notas pessoais sobre a sessão
        deviceModel:
          type: string
          nullable: true
          example: iPhone 13
          description: Modelo do dispositivo utilizado
        headphonesUsed:
          type: boolean
          example: true
          description: Se foram utilizados fones de ouvido
        roomDark:
          type: boolean
          example: false
          description: Se o ambiente estava escuro
        createdAt:
          type: string
          format: date-time
          example: 2025-08-10T14:30:00Z
          description: Data de criação do registro
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
          description: Descrição do erro
        code:
          type: string
          example: INVALID_INPUT
          description: Código do erro (opcional)
        details:
          type: object
          description: Detalhes adicionais do erro (opcional)
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Dados de entrada inválidos
            code: INVALID_INPUT
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Token inválido ou expirado
            code: UNAUTHORIZED
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso não encontrado
            code: NOT_FOUND
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Erro interno do servidor
            code: INTERNAL_ERROR
...